% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20023R.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of Cpp and R.
}
\examples{
\dontrun{
N <- 10000
x_0 <- 0
sigma <- 1
ts <- microbenchmark(rw.MetropolisR = rw.Metropolis(sigma,x_0,N), 
rw.MetropolisCpp = rw_MetropolisC(sigma,x_0,N))
summary(ts)[,c(1,3,5,6)]
}
}
