% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20023R.R
\name{geo_anomaly}
\alias{geo_anomaly}
\title{An anomaly detection function}
\usage{
geo_anomaly(data, K, b_threshold, Rmse_threshold)
}
\arguments{
\item{data}{the data matrix, which should provides device-id for its first column, latitude and longitude with colnames "lat" and "lon" for the second and third column and data in order of time in following columns}

\item{K}{the number of multi-resolution spline basis functions used}

\item{b_threshold}{the threshold used in controling the number of devices found in terms of relatively high and low measurements}

\item{Rmse_threshold}{the threshold used in controling the number of devices found in terms of malfunctioned devices}
}
\value{
A list consisting of devices with relatively high measurements, devices with relatively low measurements and malfunctioned devices, in order.
}
\description{
An anomaly detection function with method proposed by Huang G. et al.
}
\examples{
\dontrun{
data(PMdata_original)
K <- 100; b <- 3; Rmse <- 4
l <- geo_anomaly(PMdata_original,K,b,Rmse)
}
}
\references{
Huang G, Chen L-J, Hwang W-H, Tzeng S, Huang H-C. Real-time PM2.5 mapping and anomaly detection from AirBoxes in Taiwan. Environmetrics. 2018;e2537. https://doi.org/10.1002/env.2537
}
