% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20023R.R
\name{l}
\alias{l}
\title{Anomaly found by applying geo_anomaly() to PMdata_original}
\description{
Anomaly found by applying geo_anomaly() to PMdata_original with K = 100, b_threshold = 3, Rmse_threshold = 4
}
\examples{
\dontrun{
data(l)
}
}
