% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20023R.R
\name{PMdata_original}
\alias{PMdata_original}
\title{The original PM 2.5 data by airboxes in Taiwan in consecutive 57 hours.}
\description{
This data is used to provide an example which the function geo-anomaly can handle.
}
\examples{
\dontrun{
data(PMdata_original)
}
}
