% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_MetropolisC}
\alias{rw_MetropolisC}
\title{A random walk Metropolis sampler}
\usage{
rw_MetropolisC(sigma, x_0, N)
}
\arguments{
\item{sigma}{the stardard deviation of the random walk}

\item{x_0}{the start point}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n} as well as the number of rejections
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
