% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20023R.R
\name{sirs}
\alias{sirs}
\title{Sequentially and iteratively reweighted squares algorithm for sparse recovery}
\usage{
sirs(A, y, a = 0.1)
}
\arguments{
\item{A}{the design matrix}

\item{y}{the vector of response}

\item{a}{the parameter in sica penalty}
}
\value{
a estimated vector of coefficients
}
\description{
SIRS algorithm was proposed by Jinchi Lv and Yingying Fan in
}
\examples{
\dontrun{
s <- 7; n<- 35; p <- 100; r <- 0.1; beta <- c(1,-0.5,0.7,-1.2,-0.9,3,0.55);
data <- simu_sirs_generation(s,n,p,r,beta)
X <- data[[1]]
y <- data[[2]]
a <- 0.5
beta_hat <- sirs(X,y,a)
}
}
\references{
Lv, Jinchi; Fan, Yingying. A unified approach to model selection and sparse recovery using regularized least squares. Ann. Statist. 37 (2009), no. 6A, 3498--3528. doi:10.1214/09-AOS683.
}
