\name{StatComp20023-package}
\alias{StatComp20023-package}
\alias{StatComp20023}
\docType{package}
\title{
  Homework R package of Li Chen 20023
}
\description{
The Package has 2 main roles: Sparse recovery in high-dimensional statistics and anomaly detection in spatial statistics. Two vignettes are available, one for regular assignments and one for my own functions.
}
\author{
Li Chen, cl1999@mail.ustc.edu.cn.

Maintainer: Li Chen <cl1999@mail.ustc.edu.cn>
}

\keyword{ package }
